server/requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
docopt==0.6.2
h11==0.14.0
idna==3.4
pipreqs==0.4.13
psycopg2-binary==2.9.6
requests==2.28.2
sniffio==1.3.0
starlette==0.26.1
urllib3==1.26.15
uvicorn==0.21.1
yarg==0.1.9
